requirements.txt
PyMuPDF
azure-ai-documentintelligence
azure-core
azure-identity
azure-search-documents
cffi
langchain
tiktoken
python-dotenv
pillow
requests
pypdf
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchainhub
openai
Flask==2.2.5
flask-cors==5.0.1
azure-ai-contentunderstanding
